+++ lvlspy/io/ensdf/__init__.py
@@ -1,7 +1,0 @@
-"""Compatibility wrapper for :mod:`lvlspy.extensions.io.ensdf`."""
-
-# pylint: disable=duplicate-code
-
-from lvlspy.extensions.io.ensdf import update_from_ensdf, write_to_ensdf
-
-__all__ = ["update_from_ensdf", "write_to_ensdf"]